+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `opentelemetry-instrumentation-richconsole`: Add support for suppressing resource information
+
 ## Version 1.38.0/0.59b0 (2025-10-16)
 
 ### Fixed